requirements.txt
Django==2.2.25
pytz==2021.3
sqlparse==0.4.2
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-mysql==3.5.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
Pillow==7.1.2
requests==2.24.0
simplejson==3.17.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9